Heavy damage handling (#9)
* HeavyDamage unmarshall implementation

Data source has a mix of both integers and strings for HeavyDamage. To handle this we've got a custom  heavyDamage type which implements the unmarshall and string interfaces

* Weapon output refactor

Refacted the weapon output to include more relevant details. Additionaly, moved around some of the damage decisions to account for the fact there's /way/ more than three damage types.